// src/lib/api/generated/model/callSlipFromReferralSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export interface CallSlipFromReferralSchema {
  destination_code?: string;
  mode?: string;
  office_only_remarks?: string;
  referral_reference: string;
  reissue_reason_code?: string;
  reissued_from_reference?: string | null;
  report_to_destination?: string;
  student_safe_instructions?: string;
  student_safe_location?: string;
}